[module] Verplaats menu naar een module
