+++ htdocs/index.php
@@ -42,7 +42,7 @@ try {
 		/** @noinspection PhpIncludeInspection */
 		require_once 'controller/' . $class . '.class.php';
 	} else {
-		$modules = array('eetplan', 'documenten', 'bibliotheek', 'mededelingen', 'peilingen', 'gesprekken', 'courant', 'forum', 'agenda', 'fotoalbum');
+		$modules = array('eetplan', 'documenten', 'bibliotheek', 'mededelingen', 'peilingen', 'gesprekken', 'courant', 'forum', 'agenda', 'fotoalbum', 'menu');
 
 		if (in_array($module, $modules)) {
 			/** @noinspection PhpIncludeInspection */
